requirements-simple.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
sqlmodel==0.0.14
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
